pyproject.toml
[project]
name = "mot-system"
version = "0.1.0"
description = "Multiple Object Tracking system with RT-DETRv2"
authors = [{name = "CVLAB Team", email = "team@cvlab.com"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.8.0",
    "numpy>=1.21.0",
    "pillow>=9.0.0",
    "ultralytics>=8.0.0",
    "transformers>=4.30.0",
    "timm>=0.9.0",
    "scipy>=1.9.0",
    "matplotlib>=3.5.0",
    "tqdm>=4.64.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=2.20.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
